@@ -24,18 +24,17 @@
     "@types/web3": "^1.2.2",
     "@web3-react/core": "^8.2.0",
     "@web3-react/injected-connector": "^6.0.7",
-    "antd": "^5.5.2",
+    "antd": "^5.11.0",
     "eslint-config-next": "^14.0.1",
     "eslint-config-standard-with-typescript": "^39.1.1",
     "ethers": "^6.3.0",
-    "framer-motion": "^10.12.10",
     "next": "14.0.1",
     "next-transpile-modules": "^10.0.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.8.0",
     "react-intl": "^6.5.1",
-    "typescript": "5.0.4",
+    "typescript": "5.2.2",
     "web3": "^4.2.2"
   },
   "devDependencies": {
