@@ -1,11 +1,12 @@
-import { supabase } from '../../lib/supabase';
+'use client';
+import { supabase } from '../../lib/supabase-client';
 import notification from 'antd/es/notification';
 import { Button, Card, Form, Input, Select } from 'antd';
 import { UserOutlined } from '@ant-design/icons';
 import { handleAuthError } from '../../components/app/shared/error_handler';
 import { SignUpForm } from '../../components/model/forms_models';
 import Title from 'antd/es/typography/Title';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 
 export default function SignUp() {
   const router = useRouter();
